moved DB seed files and refactored to add missing fields
